requirements.txt
spotipy
pandas
numpy
scikit-learn
streamlit
python-dotenv